fix: column 모달 빌드 오류 수정
