+++ src/app/@modal/(.)dashboard-setting/edit/column-modal/page.tsx
@@ -1,25 +1,10 @@
-'use client';
-export const dynamic = 'force-dynamic';
-import React from 'react';
-import { useRouter, useSearchParams } from 'next/navigation';
-
-import EditColumn from '@/components/Modals/domains/Column/EditColumn'; 
-import modalStyles from '@/components/Modals/Modal.module.css';
+import { Suspense } from 'react';
+import ColumnModalClient from './ColumnModalClient';
 
 export default function ColumnModalPage() {
-  const router = useRouter();
-  const searchParams = useSearchParams();
-
-  const initialTitle = searchParams.get('title') ?? '';
-
-  const onClose = () => router.back();
-
   return (
-
-    <div className={`${modalStyles.overlay} ${modalStyles.open}`} onClick={onClose}>
-      <div className={modalStyles.modal} onClick={(e) => e.stopPropagation()}>
-        <EditColumn initialTitle={initialTitle} />
-      </div>
-    </div>
+    <Suspense fallback={null}>
+      <ColumnModalClient />
+    </Suspense>
   );
 }
